Renamed map components
